fix:Simplify testing
