+++ flink-connector-obkv-hbase/src/test/java/com/oceanbase/connector/flink/OBKVHBaseConnectorITCase.java
@@ -22,7 +22,6 @@ import org.apache.flink.table.api.bridge.java.StreamTableEnvironment;
 
 import org.apache.hadoop.hbase.util.Bytes;
 import org.junit.Test;
-import org.junit.jupiter.api.AfterEach;
 
 import java.sql.SQLException;
 import java.util.Arrays;
@@ -42,13 +41,6 @@ public class OBKVHBaseConnectorITCase extends OceanBaseMySQLTestBase {
         options.put("password", getPassword());
         options.put("schema-name", getSchemaName());
         return options;
-    }
-
-    @AfterEach
-    public void tearDown() {
-        if (CONTAINER != null) {
-            CONTAINER.close();
-        }
     }
 
     @Test
